requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
kafka~=1.3.5
tweepy~=4.6.0
python-dotenv
pyspark~=3.2.1